build: Support configuration in buildroot (#145)
* build: Support configuration in buildroot

Some users may not want compressed build images, and the compression
processes take quite a bit of time. This patch adds configuration
support to buildroot.sh, allowing the user to decide which (if any)
compressed build images will be produced. If the user disables both
compression formats, the uncompressed image will be left in the
build directory.

* build: Address ShellCheck warnings