installer: assume that minimum release being installed is 'Jessie' (#179)
* Remove remaining tests for Wheezy release

Most support for the Debian 'Wheezy' release was already removed,
but a few tests remain. This patch removes them.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>

* Use proper test for type of init system

Instead of assuming systemd will be the init system unless the configuration
replaces the cdebootstrap command completely, check the cdebootstrap command
and the configuration variables which list packages to see if any of them
will install the sysvinit-core package. If that package is installed then
the init system will be sysvinit, otherwise it will be systemd.

Signed-off-by: Kevin P. Fleming <kevin@km6g.us>