Add support for 64-bit arm64 architecture.
The installer userspace remains 32-bit. 64-bit kernel is booted on
64-bit CPUs so it can run 64-bit binaries when installing an arm64
system.

32-bit system comes from Raspbian while 64-bit comes from Debian, so
there are two default mirrors now, depending on arch.
This also means that Debian keys must be included in the installer.
The sources.list file of the installed system must reflect this too so
remove the static one and generate it during install (using the current
mirror).